requirements.txt
asgiref==3.6.0
astroid==2.15.4
colorama==0.4.6
dill==0.3.6
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
platformdirs==3.5.0
pylint==2.17.3
pylint-django==2.5.3
pylint-plugin-utils==0.7
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.5.0
tzdata==2023.3
wrapt==1.15.0
